requirements_2023-07-24.txt
astropy==5.3.1
Cartopy==0.21.1
certifi==2023.5.7
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.0
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
pathlib==1.0.1
Pillow==10.0.0
pyerfa==2.0.0.3
pyparsing==3.0.9
pyproj==3.6.0
pyshp==2.3.1
python-dateutil==2.8.2
PyYAML==6.0.1
shapely==2.0.1
six==1.16.0
